// src/models/interfaces.ts
export interface ITrainRoute {
	have_first_class: boolean;
	have_second_class: boolean;
	have_third_class: boolean;
	have_fourth_class: boolean;
	have_wifi: boolean;
	have_air_conditioning: boolean;
	is_express: boolean;
	min_price: number;
	available_seats: number;
	available_seats_info: IAvailableSeatsInfo;
}

export interface IItem extends ITrainRoute {
	departure: IDeparture;
	returnRoute?: IItem; // Cambiar a IItem
	arrival?: IDeparture; // Asegurarse de incluir esta propiedad opcional
}

export interface IDeparture extends ITrainRoute {
	_id: string;
	duration: number;
	train: IIdName;
	from: IFrom;
	to: IFrom;
	price_info: IPriceInfo;
}

export interface IAvailableSeatsInfo {
	first?: number;
	second?: number;
	third?: number;
	fourth?: number;
}

export interface IFrom {
	railway_station_name: string;
	city: IIdName;
	datetime: number;
}

export interface IIdName {
	_id: string;
	name: string;
}

export interface IPriceInfo {
	first?: IPriceClass;
	second?: IPriceClass;
	third?: IPriceClass;
	fourth?: IPriceClass;
}
// ...
export interface ISeats {
	coach: ICoach;
	seats: ISeatAvailable[];
}

export interface ICoach extends ITrainRoute {
	_id: string;
	name: string;
	class_type: string;
	price: number;
	top_price: number;
	bottom_price: number;
	side_price: number;
	linens_price: number;
	wifi_price: number;
	air_conditioning_price: number; // ESTO LO AGREGUE YO
	is_linens_included: boolean;
	train: string;
	seats: string;
}

export interface ISeatAvailable {
	index: number;
	available: boolean;
}